multi: implement pay per share.
This implements the pay per share payout scheme.
The project structure and packages have also
reworked. A payment method paramter has been
added to the config as well.
