+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/davecgh/go-spew v1.1.1
 	github.com/decred/dcrd/blockchain v1.0.2
 	github.com/decred/dcrd/chaincfg v1.2.0
+	github.com/decred/dcrd/chaincfg/chainhash v1.0.1
 	github.com/decred/dcrd/dcrutil v1.1.1
 	github.com/decred/dcrd/rpcclient v1.0.2
 	github.com/decred/dcrd/wire v1.2.0
