--- main.go
@@ -1,53 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"os"
-	"os/signal"
-	"runtime"
-)
-
-var cfg *config
-var pool *MiningPool
-
-func main() {
-	// Use all processor cores.
-	runtime.GOMAXPROCS(runtime.NumCPU())
-
-	// Listen for interrupt signals.
-	interrupt := make(chan os.Signal, 1)
-	signal.Notify(interrupt, os.Interrupt)
-
-	// Load configuration and parse command line. This also initializes logging
-	// and configures it accordingly.
-	cfg, _, err := loadConfig()
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	defer func() {
-		if logRotator != nil {
-			logRotator.Close()
-		}
-	}()
-
-	p, err := NewMiningPool(cfg)
-	if err != nil {
-		pLog.Error(err)
-		return
-	}
-
-	pLog.Infof("Version: %s", version())
-	pLog.Infof("Runtime: Go version %s", runtime.Version())
-	pLog.Infof("Home dir: %s", cfg.HomeDir)
-	pLog.Infof("Started dcrpool.")
-
-	go p.listen()
-	for {
-		select {
-		case <-interrupt:
-			go p.shutdown()
-			return
-		}
-	}
-}
